Allow for warning-free build of ca65 with MS VC on warning level 3.

+++ src/ca65/expr.c
@@ -1031,7 +1031,7 @@ static ExprNode* Factor (void)
 	    break;
 
 	case TOK_TIME:
-	    N = GenLiteralExpr (time (0));
+	    N = GenLiteralExpr ((long) time (0));
 	    NextTok ();
 	    break;
