+++ src/ca65/scanner.c
@@ -526,7 +526,7 @@ int NewInputFile (const char* Name)
     /* Add the file to the input file table and remember the index */
     FileIdx = AddFile (SB_InitFromString (&NameBuf, Name),
                        (FCount == 0)? FT_MAIN : FT_INCLUDE,
-                       Buf.st_size, Buf.st_mtime);
+                       Buf.st_size, (unsigned long) Buf.st_mtime);
 
     /* Create a new input source variable and initialize it */
     S                   = xmalloc (sizeof (*S));
